Remove Ether Renewal Bonder class and refactor Ether Renewal Prot Infuser to incorporate its functionality
